📦 NEW: features,bugfixes,better approaches,organized work

+++ database/migrations/2020_03_22_170607_create_bills_table.php
@@ -17,7 +17,7 @@ class CreateBillsTable extends Migration
             $table->Increments('id');
             $table->unsignedInteger('user_id');
                 $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedInteger('weight');
+            $table->float('weight');
             $table->unsignedInteger('project_id');
                 $table->foreign('project_id')->references('id')->on('projects');
             $table->timestamps();
